+++ src/z80asm/ldinstr.c
@@ -13,7 +13,7 @@
 Copyright (C) Gunther Strube, InterLogic 1993-99
 */
 
-/* $Header: /home/dom/z88dk-git/cvs/z88dk/src/z80asm/Attic/ldinstr.c,v 1.3 2009-07-18 23:23:15 dom Exp $ */
+/* $Header: /home/dom/z88dk-git/cvs/z88dk/src/z80asm/Attic/ldinstr.c,v 1.4 2010-04-16 17:34:37 dom Exp $ */
 /* $History: LDINSTR.C $ */
 /*  */
 /* *****************  Version 10  ***************** */
@@ -52,7 +52,7 @@ Copyright (C) Gunther Strube, InterLogic 1993-99
 
 /* external functions */
 enum symbols GetSym (void);
-void ReportError (char *filename, short linenr, int errnum);
+void ReportError (char *filename, int linenr, int errnum);
 void RemovePfixlist (struct expr *pfixexpr);
 int ExprUnsigned8 (int listoffset);
 int ExprSigned8 (int listoffset);
